feat(server): working on updating the database
Trying to update the database based on the status code returned from git
(A|D|M). I also created some helper functions for interacting with the
note model with MongoDB. Where I put it might be the wrong place. I'm
not sure. I still need to do some major testing to be sure that
everything is working.