feat(server): add `/getNoteByBlobHash` endpoint
the `/getNoteByBlobHash` endpoint will return a note based on the blob
hash supplied. I added express-validator to check that the body of the
request makes sense and is in the proper format.

I need to make a proper logging helper class to work with logging to the
console and throwing errors back from express.