feat(client,server): fix server and starting on note page
starting on the individual notes page by sending a param to the page we
want to view. If it doesn't exist we are going to have to request it. I
don't know how this is going to work yet since it's only using the title
string but we will see.